Post tagging MVC delivery, update start

+++ jpa/src/main/java/com/nixmash/springdata/jpa/dto/TagDTO.java
@@ -9,7 +9,7 @@ public class TagDTO implements Serializable {
 
     private static final long serialVersionUID = -4809849404139121173L;
 
-    private Integer tagId = -1;
+    private long tagId = -1;
     private String tagValue;
 
     public TagDTO() {
@@ -20,11 +20,11 @@ public class TagDTO implements Serializable {
     }
 
 
-    public Integer getTagId() {
+    public long getTagId() {
         return tagId;
     }
 
-    public void setTagId(Integer tagId) {
+    public void setTagId(long tagId) {
         this.tagId = tagId;
     }
 
@@ -36,9 +36,10 @@ public class TagDTO implements Serializable {
         this.tagValue = tagValue;
     }
 
-    public TagDTO(Integer tagId, String tagValue) {
+    public TagDTO(long tagId, String tagValue) {
         this.tagId = tagId;
         this.tagValue = tagValue;
     }
 
 }
+
